Add Error de conexion a la BDD.

+++ index.php
@@ -2,13 +2,18 @@
 session_start();
 require_once("database/Database.php");
 $estaLogeado = false;
+$errorBdd = "";
 
 if (isset($_SESSION['usuario'])){
     $estaLogeado = true;
 }
 
 $bdd = new Database();
-$pokemones = $bdd->getPorQuery("SELECT * FROM `pokemon` INNER JOIN `tipo` ON `pokemon`.`id_tipo` = `tipo`.`id_tipo`");
+if ($bdd->getError() != ""){
+    $errorBdd = $bdd->getError();
+}else{
+    $pokemones = $bdd->getPorQuery("SELECT * FROM `pokemon` INNER JOIN `tipo` ON `pokemon`.`id_tipo` = `tipo`.`id_tipo`");
+}
 //echo json_encode($pokemones);
 ?>
 
@@ -18,6 +23,8 @@ $pokemones = $bdd->getPorQuery("SELECT * FROM `pokemon` INNER JOIN `tipo` ON `po
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css" rel="stylesheet">
     <link rel="stylesheet" href="/TP-Pokedex/styles/home.css">
     <title>Inicio | Pokedex</title>
 </head>
@@ -33,28 +40,32 @@ $pokemones = $bdd->getPorQuery("SELECT * FROM `pokemon` INNER JOIN `tipo` ON `po
 
 
         <section class="contenedor_pokemones">
-            <?php foreach ($pokemones as $pokemon){ ?>
-                <div class="pokemon">
-                    <a href="/TP-Pokedex/views/detallePokemon.php?id_pokemon=<?php echo $pokemon["id_pokemon"]?>" class="cont_1">
-                        <img src="<?php echo $pokemon["imagen"]?>" alt="img_pokemon">
-                    </a>
-                    <div class="cont_2">
-                        <a href="/TP-Pokedex/views/detallePokemon.php?id_pokemon=<?php echo $pokemon["id_pokemon"]?>" class="descripcion">
-                            <p class="numero_pokemon"><?php echo $pokemon["id_pokemon"]?></p>
-                            <p class="nombre_pokemon"><?php echo $pokemon["nombre"]?></p>
-                            <img src="/TP-Pokedex/assets/tipos/<?php echo $pokemon["nombre_tipo"]?>.avif" alt="img_tipo">
+            <?php
+            if ($errorBdd != ""){
+                echo '<div class="alert alert-danger">' . $errorBdd . '</div>';
+            }else{
+                foreach ($pokemones as $pokemon){ ?>
+                    <div class="pokemon">
+                        <a href="/TP-Pokedex/views/detallePokemon.php?id_pokemon=<?php echo $pokemon["id_pokemon"]?>" class="cont_1">
+                            <img src="<?php echo $pokemon["imagen"]?>" alt="img_pokemon">
                         </a>
-                        <?php
+                        <div class="cont_2">
+                            <a href="/TP-Pokedex/views/detallePokemon.php?id_pokemon=<?php echo $pokemon["id_pokemon"]?>" class="descripcion">
+                                <p class="numero_pokemon"><?php echo $pokemon["id_pokemon"]?></p>
+                                <p class="nombre_pokemon"><?php echo $pokemon["nombre"]?></p>
+                                <img src="/TP-Pokedex/assets/tipos/<?php echo $pokemon["nombre_tipo"]?>.avif" alt="img_tipo">
+                            </a>
+                            <?php
 
-                        if($estaLogeado){?>
-                            <div class="cont_botones">
-                                <a href="editar" class="btn_editar"> <img src="assets/icons/icon_edit.svg">Editar</a>
-                                <a href="eliminar" class="btn_eliminar"> <img src="assets/icons/icon_delete.svg">Eliminar</a>
-                            </div>
-                        <?php }?>
+                            if($estaLogeado){?>
+                                <div class="cont_botones">
+                                    <a href="editar" class="btn_editar"> <img src="assets/icons/icon_edit.svg">Editar</a>
+                                    <a href="eliminar" class="btn_eliminar"> <img src="assets/icons/icon_delete.svg">Eliminar</a>
+                                </div>
+                            <?php }?>
+                        </div>
                     </div>
-                </div>
-            <?php } ?>
+                <?php }} ?>
         </section>
 
         <div class="container_ancla">
@@ -67,5 +78,9 @@ $pokemones = $bdd->getPorQuery("SELECT * FROM `pokemon` INNER JOIN `tipo` ON `po
 
 
     <?php require_once("views/footer.php")?>
+
+
+    <!-- Bootstrap JS -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
 </body>
 </html>
